tests: separate envtest directory and makefile target
