+++ .github/workflows/_envtest_tests.yaml
@@ -1,0 +1,36 @@
+name: envtest tests
+
+on:
+  workflow_call: {}
+
+jobs:
+  envtest-tests:
+    runs-on: ubuntu-latest
+    steps:
+      - name: checkout repository
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
+      - name: setup golang
+        uses: actions/setup-go@v4
+        with:
+          go-version: '^1.20'
+
+      - name: run envtest tests
+        run: make test.envtest
+        env:
+          GOTESTSUM_JUNITFILE: envtest-tests.xml
+
+      - name: collect test coverage
+        uses: actions/upload-artifact@v3
+        with:
+          name: coverage
+          path: coverage.envtest.out
+
+      - name: collect test report
+        if: ${{ always() }}
+        uses: actions/upload-artifact@v3
+        with:
+          name: tests-report
+          path: envtest-tests.xml
